Fix exact/inexact coding
The idea is that encode-f32 is called for
non integers, including 1/2 etc. so it must
coerce its input to inexact, likewise encode-i32
is called for 40.0, so must coerce to exact.
The important fix is that the type tag and
the encoding choice _must_ be selected on
the same rule (ie. integer? -> i32 else f32)

darcs-hash:ea80f0ca56fd252a94d07ea702bf9bf969dd3384